launcherFinalizer
Move launcherFinalizer into it’s own file. Create a window variant.
